problems/src/test/java/_1051_1100/_1085_Sum_of_Digits_in_the_Minimum_Number/SolutionITest.java
package _1051_1100._1085_Sum_of_Digits_in_the_Minimum_Number;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

/**
 * @author jacka
 * @version 1.0 on 6/27/2021
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase1(Solution solution) {
    final int[] nums = {34, 23, 1, 24, 75, 33, 54, 8};
    assertEquals(0, solution.sumOfDigits(nums));
  }

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase2(Solution solution) {
    final int[] nums = {99, 77, 33, 66, 55};
    assertEquals(1, solution.sumOfDigits(nums));
  }

  static Stream<Solution> solutionStream() {
    return Stream.of(
        new SolutionI()
    );
  }
}